+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Update deltachat-core-rust to 1.46.0
 - BREAKING: There is no `serverFlags` option anymore, use `mail_security` and `send_security`
 - BREAKING: Remove `DCN_EVENT_CONFIGURE_FAILED` and `DCN_EVENT_CONFIGURE_SUCCESSFUL`
+- BREAKING: Removed deprecated dcn_chat_get_archived and dcn_archive_chat
+
 
 ## [1.45.0] - 2020-03-10